+++ src/lib/pages/public-project/slug.tsx
@@ -89,7 +89,7 @@ export const ProjectDetail = observer(() => {
                 borderTop="1px solid"
                 borderColor="pebble.700"
               >
-                <EmptyState message="There is currently no code in this project. Please check back soon for the updates." />
+                <EmptyState message="There is currently no code related to this project." />
               </Flex>
             )}
             <Heading as="h6" variant="h6" mb={6} mt={12} px="48px">
@@ -128,7 +128,7 @@ export const ProjectDetail = observer(() => {
                 borderTop="1px solid"
                 borderColor="pebble.700"
               >
-                <EmptyState message="There is currently no contracts in this project. Please check back soon for the updates." />
+                <EmptyState message="There is currently no contracts related to this project." />
               </Flex>
             )}
           </TabPanel>